+++ tests/metro/js/metro.js
@@ -1,7 +1,7 @@
 /*
  * Metro 4 Components Library v4.4.3  (https://metroui.org.ua)
  * Copyright 2012-2020 Sergey Pimenov
- * Built at 09/11/2020 11:50:07
+ * Built at 11/11/2020 15:10:57
  * Licensed under MIT
  */
 (function (global, undefined) {
@@ -313,7 +313,7 @@ function isTouch() {
         return;
     }
 
-    // 
+    // console.log("Promise polyfill v1.2.0");
 
     var PENDING = 'pending';
     var SEALED = 'sealed';
@@ -2310,7 +2310,6 @@ $.fn.extend({
 
     scrollTop: function(val){
         if (not(val)) {
-            
             return this.length === 0 ? undefined : this[0] === window ? pageYOffset : this[0].scrollTop;
         }
         return this.each(function(){
@@ -2689,7 +2688,7 @@ $.fn.extend({
                 });
             } else {
                 el.setAttribute(name, val);
-                // 
+                // console.log(name, val);
             }
         });
     },
@@ -4537,7 +4536,7 @@ $.noConflict = function() {
     var Metro = {
 
         version: "4.4.3",
-        compileTime: "09/11/2020 11:50:07",
+        compileTime: "11/11/2020 15:10:57",
         buildNumber: "@@build",
         isTouchable: isTouch,
         fullScreenEnabled: document.fullscreenEnabled,
@@ -4804,7 +4803,7 @@ $.noConflict = function() {
                         }
 
                     } else  {
-                        //
+                        //console.log(mutation);
                     }
                 });
             };
@@ -5034,6 +5033,10 @@ $.noConflict = function() {
             return Metro.$()($(el)[0]);
         },
 
+        get$elements: function(el){
+            return Metro.$()($(el));
+        },
+
         getPlugin: function(el, name){
             var _name = normalizeComponentName(name);
             var $el = Metro.get$el(el);
@@ -5042,7 +5045,7 @@ $.noConflict = function() {
 
         makePlugin: function(el, name, options){
             var _name = normalizeComponentName(name);
-            var $el = Metro.get$el(el);
+            var $el = Metro.get$elements(el);
             return $el.length && typeof $el[_name] === "function" ? $el[_name](options) : undefined;
         },
 
@@ -11557,6 +11560,430 @@ $.noConflict = function() {
         }
     });
 }(Metro, m4q));
+
+/* eslint-disable */
+(function(Metro, $) {
+    'use strict';
+
+    var Utils = Metro.utils;
+    var ColorSelectorDefaultConfig = {
+        defaultSwatches: "#FFFFFF,#000000,#FFFB0D,#0532FF,#FF9300,#00F91A,#FF2700,#686868,#EE5464,#D27AEE,#5BA8C4,#E64AA9,#1ba1e2,#6a00ff,#bebebe,#f8f8f8",
+        returnValueType: "hex",
+        returnAsString: true,
+        showValues: "hex, rgb, hsl, hsv, cmyk",
+        clsSelector: "",
+        clsSwatches: "",
+        clsSwatch: "",
+        clsValue: "",
+        clsLabel: "",
+        clsInput: "",
+        onColor: Metro.noop,
+        onMyObjectCreate: Metro.noop
+    };
+
+    Metro.colorSelectorSetup = function (options) {
+        ColorSelectorDefaultConfig = $.extend({}, ColorSelectorDefaultConfig, options);
+    };
+
+    if (typeof window["metroColorSelectorSetup"] !== undefined) {
+        Metro.colorSelectorSetup(window["metroColorSelectorSetup"]);
+    }
+
+    Metro.Component('color-selector', {
+        init: function( options, elem ) {
+            this._super(elem, options, ColorSelectorDefaultConfig, {
+                // define instance vars here
+                id: Utils.elementId("color-selector"),
+                defaultSwatches: [],
+                showValues: [],
+                hue: 0,
+                saturation: 0,
+                lightness: 1,
+                hsl: null,
+                hsv: null,
+                rgb: null,
+                cmyk: null,
+                hex: null
+            });
+            return this;
+        },
+
+        _create: function(){
+            var that = this, element = this.element, o = this.options;
+
+            this.defaultSwatches = o.defaultSwatches.toArray(",");
+            this.showValues = o.showValues.toArray(",");
+
+            this._createStructure();
+            this._createEvents();
+
+            this._fireEvent('color-selector-create');
+        },
+
+        _createStructure: function(){
+            var that = this, element = this.element, o = this.options;
+            var colorBox, row, swatches, map, value, inputs, radios;
+
+            element.addClass("color-selector").addClass(o.clsSelector);
+
+            element.append( colorBox = $("<div>").addClass("color-box") );
+
+            colorBox.append( row = $("<div>").addClass("row") );
+
+            row.append( swatches = $("<div>").addClass("default-swatches").addClass(o.clsSwatches) );
+            $.each(this.defaultSwatches, function(){
+                swatches.append(
+                    $("<button>")
+                        .attr("data-color", this)
+                        .attr("type", "button")
+                        .addClass("swatch")
+                        .addClass(o.clsSwatch)
+                        .css("background-color", this)
+                );
+            });
+
+            colorBox.append( row = $("<div>").addClass("row") );
+
+            row.append( map = $("<div>").addClass("color-map") );
+            map.append( $("<button>").attr("type", "button").addClass("cursor color-cursor") )
+            map.append( $("<canvas>").addClass("color-canvas") )
+
+            row.append( map = $("<div>").addClass("hue-map") );
+            map.append( $("<button>").attr("type", "button").addClass("cursor hue-cursor") )
+            map.append( $("<canvas>").addClass("hue-canvas") )
+
+            colorBox.append( row = $("<div>").addClass("row") );
+
+            row.append( value = $("<div>").addClass("color-value-hex") );
+            value.append( $("<input type='radio' name='returnType' value='hex' checked>").addClass("check-color-value-hex") );
+            value.append( $("<input type='text' data-prepend='HEX:' readonly>").addClass("input-small color-value-x") );
+
+            row.append( value = $("<div>").addClass("color-value-rgb") );
+            value.append( $("<input type='radio' name='returnType' value='rgb'>").addClass("check-color-value-rgb") );
+            value.append( $("<input type='text' data-prepend='R:' readonly>").addClass("input-small color-value-r") );
+            value.append( $("<input type='text' data-prepend='G:' readonly>").addClass("input-small color-value-g") );
+            value.append( $("<input type='text' data-prepend='B:' readonly>").addClass("input-small color-value-b") );
+
+            row.append( value = $("<div>").addClass("color-value-hsl") );
+            value.append( $("<input type='radio' name='returnType' value='hsl'>").addClass("check-color-value-hsv") );
+            value.append( $("<input type='text' data-prepend='H:' readonly>").addClass("input-small color-value-h") );
+            value.append( $("<input type='text' data-prepend='S:' readonly>").addClass("input-small color-value-s") );
+            value.append( $("<input type='text' data-prepend='L:' readonly>").addClass("input-small color-value-l") );
+
+            row.append( value = $("<div>").addClass("color-value-hsv") );
+            value.append( $("<input type='radio' name='returnType' value='hsv'>").addClass("check-color-value-hsl") );
+            value.append( $("<input type='text' data-prepend='H:' readonly>").addClass("input-small color-value-h") );
+            value.append( $("<input type='text' data-prepend='S:' readonly>").addClass("input-small color-value-s") );
+            value.append( $("<input type='text' data-prepend='V:' readonly>").addClass("input-small color-value-v") );
+
+            row.append( value = $("<div>").addClass("color-value-cmyk") );
+            value.append( $("<input type='radio' name='returnType' value='cmyk'>").addClass("check-color-value-cmyk") );
+            value.append( $("<input type='text' data-prepend='C:' readonly>").addClass("input-small color-value-c") );
+            value.append( $("<input type='text' data-prepend='M:' readonly>").addClass("input-small color-value-m") );
+            value.append( $("<input type='text' data-prepend='Y:' readonly>").addClass("input-small color-value-y") );
+            value.append( $("<input type='text' data-prepend='K:' readonly>").addClass("input-small color-value-k") );
+
+            inputs = colorBox.find("input[type=text]");
+            Metro.makePlugin(inputs, 'input', {
+                clearButton: false,
+                clsPrepend: o.clsLabel,
+                clsComponent: o.clsInput
+            });
+            inputs.addClass(o.clsValue);
+
+            radios = colorBox.find("input[type=radio]");
+            radios.each(function(){
+                if ($(this).val() === o.returnValueType) {
+                    this.checked = true;
+                }
+            });
+            Metro.makePlugin(radios, 'radio', {
+                style: 2
+            });
+
+            $.each(["hex", "rgb", "hsv", "hsl", "cmyk"], function(){
+                if (that.showValues.indexOf(this) === -1) element.find(".color-value-"+this).hide();
+            });
+
+            this.hueCanvas = element.find(".hue-canvas");
+            this.hueCursor = element.find(".hue-cursor");
+            this.shadeCanvas = element.find(".color-canvas");
+            this.shadeCursor = element.find(".color-cursor");
+
+            this._createShadeCanvas();
+            this._createHueCanvas();
+            this._setColorValues();
+            this._updateCursorsColor();
+        },
+
+        _createShadeCanvas: function(color){
+            var canvas = this.shadeCanvas[0];
+            var ctx = canvas.getContext('2d');
+
+            ctx.clearRect(0, 0, canvas.width, canvas.height);
+
+            if(!color) color = '#f00';
+
+            ctx.fillStyle = color;
+            ctx.fillRect(0, 0, canvas.width, canvas.height);
+
+            var whiteGradient = ctx.createLinearGradient(0, 0, canvas.width, 0);
+            whiteGradient.addColorStop(0, "#fff");
+            whiteGradient.addColorStop(1, "transparent");
+            ctx.fillStyle = whiteGradient;
+            ctx.fillRect(0, 0, canvas.width, canvas.height);
+
+            var blackGradient = ctx.createLinearGradient(0, 0, 0, canvas.height);
+            blackGradient.addColorStop(0, "transparent");
+            blackGradient.addColorStop(1, "#000");
+            ctx.fillStyle = blackGradient;
+            ctx.fillRect(0, 0, canvas.width, canvas.height);
+        },
+
+        _createHueCanvas: function(){
+            var canvas = this.hueCanvas[0];
+            var ctx = canvas.getContext('2d');
+            var hueGradient = ctx.createLinearGradient(0, 0, 0, canvas.height);
+
+            hueGradient.addColorStop(0.00, "hsl(0,100%,50%)");
+            hueGradient.addColorStop(0.17, "hsl(298.8, 100%, 50%)");
+            hueGradient.addColorStop(0.33, "hsl(241.2, 100%, 50%)");
+            hueGradient.addColorStop(0.50, "hsl(180, 100%, 50%)");
+            hueGradient.addColorStop(0.67, "hsl(118.8, 100%, 50%)");
+            hueGradient.addColorStop(0.83, "hsl(61.2,100%,50%)");
+            hueGradient.addColorStop(1.00, "hsl(360,100%,50%)");
+            ctx.fillStyle = hueGradient;
+            ctx.fillRect(0, 0, canvas.width, canvas.height);
+        },
+
+        _updateHueCursor: function(y){
+            this.hueCursor.css({
+                "top": y
+            });
+        },
+
+        _getHueColor: function(pageY){
+            var canvas = this.hueCanvas[0];
+            var rect = canvas.getBoundingClientRect();
+            var y, percent, color, hue;
+
+            y = pageY - rect.top;
+
+            if ( y > rect.height ) y = rect.height;
+            if ( y < 0 ) y = 0;
+
+            percent = y / rect.height;
+            hue = 360 - (360 * percent);
+            if (hue === 360) hue = 0;
+            color = "hsl("+ hue +", 100%, 50%)";
+            this.hue = hue;
+
+            this._createShadeCanvas(color);
+            this._updateHueCursor(y);
+            this._updateCursorsColor();
+            this._setColorValues();
+        },
+
+        _getShadeColor: function(pageX, pageY){
+            var colorRect = this.shadeCanvas[0].getBoundingClientRect();
+            var x = pageX - colorRect.left;
+            var y = pageY - colorRect.top;
+
+            if(x > colorRect.width) x = colorRect.width;
+            if(x < 0) x = 0;
+            if(y > colorRect.height) y = colorRect.height;
+            if(y < 0) y = .1;
+
+            var xRatio = x / colorRect.width * 100;
+            var yRatio = y / colorRect.height * 100;
+            var hsvValue = 1 - (yRatio / 100);
+            var hsvSaturation = xRatio / 100;
+            var lightness = (hsvValue / 2) * (2 - hsvSaturation);
+            var saturation = (hsvValue * hsvSaturation) / (1 - Math.abs(2 * lightness - 1));
+
+            if (isNaN(lightness)) {
+                lightness = 0;
+            }
+
+            if (isNaN(saturation)) {
+                saturation = 0;
+            }
+
+            this.lightness = (Math.round(lightness * 100) / 100).toFixed(1);
+            this.saturation = (Math.round(saturation * 100) / 100).toFixed(1);
+
+            this._updateColorCursor(x, y);
+            this._updateCursorsColor();
+            this._setColorValues();
+        },
+
+        _updateCursorsColor: function(){
+            this.shadeCursor.css({backgroundColor: Metro.colors.toHEX(new Metro.colorPrimitive.HSL(this.hue, this.saturation, this.lightness))});
+            this.hueCursor.css({backgroundColor: Metro.colors.toHEX(new Metro.colorPrimitive.HSL(this.hue, 1, .5))});
+        },
+
+        _updateColorCursor: function(x, y){
+            this.shadeCursor.css({
+                top: y,
+                left: x
+            })
+        },
+
+        _colorToPos: function (color){
+            var shadeCanvasRect = this.shadeCanvas[0].getBoundingClientRect();
+            var hueCanvasRect = this.hueCanvas[0].getBoundingClientRect();
+            var hsl = Metro.colors.toHSL(color);
+            var hsv = Metro.colors.toHSV(color);
+            var x = shadeCanvasRect.width * hsv.s;
+            var y = shadeCanvasRect.height * (1 - hsv.v);
+            var hueY = hueCanvasRect.height - ((hsl.h / 360) * hueCanvasRect.height);
+
+            this.hue = hsl.h;
+            this.saturation = hsl.s;
+            this.lightness = hsl.l;
+
+            this._updateHueCursor(hueY);
+            this._updateColorCursor(x, y);
+            this._updateCursorsColor();
+            this._createShadeCanvas("hsl("+ this.hue +", 100%, 50%)");
+            this._setColorValues();
+        },
+
+        _setColorValues: function(){
+            var element = this.element;
+            var hsl = Metro.colors.toHSL(new Metro.colorPrimitive.HSL(this.hue, this.saturation, this.lightness));
+            var rgb = Metro.colors.toRGB(hsl);
+            var hsv = Metro.colors.toHSV(hsl);
+            var cmyk = Metro.colors.toCMYK(hsl);
+            var hex = Metro.colors.toHEX(hsl);
+
+            this.hsl = hsl;
+            this.hsv = hsv;
+            this.rgb = rgb;
+            this.hex = hex;
+            this.cmyk = cmyk;
+
+            element.find(".color-value-hex .color-value-x input").val(hex);
+
+            element.find(".color-value-rgb .color-value-r input").val(rgb.r);
+            element.find(".color-value-rgb .color-value-g input").val(rgb.g);
+            element.find(".color-value-rgb .color-value-b input").val(rgb.b);
+
+            element.find(".color-value-hsl .color-value-h input").val(hsl.h.toFixed(0));
+            element.find(".color-value-hsl .color-value-s input").val(hsl.s.toFixed(4));
+            element.find(".color-value-hsl .color-value-l input").val(hsl.l.toFixed(4));
+
+            element.find(".color-value-hsv .color-value-h input").val(hsv.h.toFixed(0));
+            element.find(".color-value-hsv .color-value-s input").val(hsv.s.toFixed(4));
+            element.find(".color-value-hsv .color-value-v input").val(hsv.v.toFixed(4));
+
+            element.find(".color-value-cmyk .color-value-c input").val(cmyk.c.toFixed(0));
+            element.find(".color-value-cmyk .color-value-m input").val(cmyk.m.toFixed(0));
+            element.find(".color-value-cmyk .color-value-y input").val(cmyk.y.toFixed(0));
+            element.find(".color-value-cmyk .color-value-k input").val(cmyk.k.toFixed(0));
+
+            this._fireEvent("color", {
+                hue: this.hue,
+                saturation: this.saturation,
+                lightness: this.lightness,
+                color: this.val()
+            });
+        },
+
+        _createEvents: function(){
+            var that = this, element = this.element, o = this.options;
+            var hueCanvas = this.hueCanvas;
+            var shadeCanvas = this.shadeCanvas;
+            var radios = element.find("input[type=radio]");
+
+            radios.on("click", function(){
+                o.returnValueType = $(this).val();
+                console.log(that.val());
+            });
+
+            hueCanvas.on("mousedown", function(e){
+                e.preventDefault();
+                that._getHueColor(e.pageY);
+                that.hueCursor.addClass("dragging");
+
+                $(window).on("mousemove", function(e){
+                    e.preventDefault();
+                    that._getHueColor(e.pageY);
+                }, {ns: that.id});
+
+                $(window).on("mouseup", function(){
+                    that.hueCursor.removeClass("dragging");
+                    $(window).off("mousemove", {ns: that.id});
+                })
+            });
+
+            shadeCanvas.on("mousedown", function(e){
+                e.preventDefault();
+                that._getShadeColor(e.pageX, e.pageY);
+                that.shadeCursor.addClass("dragging");
+
+                $(window).on("mousemove", function(e){
+                    e.preventDefault();
+                    that._getShadeColor(e.pageX, e.pageY);
+                }, {ns: that.id});
+
+                $(window).on("mouseup", function(){
+                    that.shadeCursor.removeClass("dragging");
+                    $(window).off("mousemove", {ns: that.id});
+                })
+            });
+
+            element.on("click", ".swatch", function(e){
+                that._colorToPos($(this).attr("data-color"));
+                e.preventDefault();
+            })
+        },
+
+        val: function(v){
+            var o = this.options;
+
+            if (!Utils.isValue(v) || !Metro.colors.isColor(v)) {
+                var res;
+                switch (o.returnValueType.toLowerCase()) {
+                    case "rgb":
+                        res = this.rgb;
+                        break;
+                    case "hsl":
+                        res = this.hsl;
+                        break;
+                    case "hsv":
+                        res = this.hsv;
+                        break;
+                    case "cmyk":
+                        res = this.cmyk;
+                        break;
+                    default: res = this.hex;
+                }
+                return o.returnAsString ? res.toString() : res;
+            }
+
+            this._colorToPos(Metro.colors.toHEX(v));
+        },
+
+        changeAttribute: function(attr, newValue){
+            var o = this.options;
+
+            if (attr === "data-return-value-type") {
+                o.returnValueType = newValue;
+            }
+
+            if (attr === "data-return-as-string") {
+                o.returnValueType = Utils.bool(newValue);
+            }
+        },
+
+        destroy: function(){
+            this.element.remove();
+        }
+    });
+}(Metro, m4q));
+/* eslint-enable */
+
 
 (function(Metro, $) {
     'use strict';